Add out directory to workaround bug in build script
This is not the "ideal" solution but I just noticed the build script was broken and wanted to make it work immediately without effort. This software is something I used several years ago on an AI project for an employer's project. I am considering refining this on my own just cause I'm bored and need to refresh my C skills.